+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.5] - 2018-11-30
 ### Added
 - `SQLiteCursor#getDateTime()` method
+- `SQLiteDatabase#isDatabaseIntegrityOk` getter
 - `SQLiteDatabase#validateSql()` method
 - `SQLiteDatabase#yieldIfContendedSafely()` method
 - `SQLiteDatatypeMismatchException` exception